+++ components/cascader/cascader.vue
@@ -19,6 +19,8 @@
                     :isOpened="isOpened"
                     :multiple="multiple"
                     :placeholder="placeholder"
+                    :collapseTags="collapseTags"
+                    :collapseTagsLimit="collapseTagsLimit"
                     @remove="handleRemove"
                     @clear="handleClear"
                 />
@@ -40,16 +42,8 @@
     </div>
 </template>
 <script>
-import {
-    defineComponent,
-    ref,
-    unref,
-    watch,
-    computed,
-} from 'vue';
-import {
-    cloneDeep,
-} from 'lodash-es';
+import { defineComponent, ref, unref, watch, computed } from 'vue';
+import { cloneDeep } from 'lodash-es';
 import getPrefixCls from '../_util/getPrefixCls';
 import { useNormalModel } from '../_util/use/useModel';
 import { UPDATE_MODEL_EVENT, CHANGE_EVENT } from '../_util/constants';
@@ -74,7 +68,7 @@ export default defineComponent({
         },
         nodeConfig: {
             type: Object,
-            default: () => { },
+            default: () => {},
         },
         placeholder: {
             type: String,
@@ -106,6 +100,14 @@ export default defineComponent({
         },
         getContainer: {
             type: Function,
+        },
+        collapseTags: {
+            type: Boolean,
+            default: false,
+        },
+        collapseTagsLimit: {
+            type: Number,
+            default: 1,
         },
     },
     emits: [
@@ -140,9 +142,15 @@ export default defineComponent({
             const { emitPath } = props.nodeConfig || {};
             const copyValue = cloneDeep(currentValue.value);
             if (emitPath) {
-                copyValue.splice(copyValue.indexOf(item => item.includes(value)), 1);
+                copyValue.splice(
+                    copyValue.indexOf((item) => item.includes(value)),
+                    1,
+                );
             } else {
-                copyValue.splice(copyValue.indexOf(item => item === value), 1);
+                copyValue.splice(
+                    copyValue.indexOf((item) => item === value),
+                    1,
+                );
             }
             updateCurrentValue(copyValue);
             emit('removeTag', value);
@@ -159,10 +167,14 @@ export default defineComponent({
         const handleUpdateSelectedNodes = (value) => {
             selectedNodes.value = value;
         };
-        const selectedOptions = computed(() => selectedNodes.value.map(selectedNode => ({
-            value: selectedNode.value,
-            label: props.showAllLevels ? selectedNode.pathLabels.join(props.separator) : selectedNode.label,
-        })));
+        const selectedOptions = computed(() =>
+            selectedNodes.value.map((selectedNode) => ({
+                value: selectedNode.value,
+                label: props.showAllLevels
+                    ? selectedNode.pathLabels.join(props.separator)
+                    : selectedNode.label,
+            })),
+        );
 
         return {
             prefixCls,