+++ components/form/formItem.vue
@@ -19,8 +19,9 @@ import {
 } from 'vue';
 import Schema from 'async-validator';
 import { isArray, cloneDeep } from 'lodash-es';
+import { addUnit } from '../_util/utils';
 import {
-    provideKey, FORM_ITEM_NAME, LABEL_POSITION, TRIGGER_DEFAULT, VALIDATE_STATUS, VALIDATE_MESSAGE_DEFAULT,
+    provideKey, FORM_ITEM_NAME, LABEL_POSITION, TRIGGER_DEFAULT, VALIDATE_STATUS, VALIDATE_MESSAGE_DEFAULT, LABEL_MARGIN_RIGHT_DEFAULT
 } from './const';
 import getPrefixCls from '../_util/getPrefixCls';
 import { FORMITEM_INJECTION_KEY } from '../_util/constants';
@@ -33,6 +34,7 @@ export default defineComponent({
     props: {
         prop: String,
         label: String,
+        labelWidth: String | Number,
         showMessage: {
             type: Boolean,
             default: null,
@@ -48,6 +50,7 @@ export default defineComponent({
             rules,
             showMessage,
             labelWidth,
+            labelMarginRight,
             labelPosition,
             addField,
             removeField,
@@ -78,8 +81,11 @@ export default defineComponent({
                 .concat((formItemRequired.value && ['is-required']) || []) // 必填校验: is-required
                 .concat((validateStatus.value === VALIDATE_STATUS.ERROR && ['is-error']) || []); // 校验错误: is-error
             return classSet.join(' ');
-        });
-        const formItemLabelStyle = computed(() => ({ width: `${props.labelWidth || labelWidth.value}px` }));
+        });        
+        const formItemLabelStyle = computed(() => ({ 
+            width: addUnit(props.labelWidth || labelWidth.value),
+            'margin-right': addUnit(labelMarginRight.value) || LABEL_MARGIN_RIGHT_DEFAULT
+        }));
 
         let ruleDefaultType = 'string';
         const setRuleDefaultType = (val) => {