+++ components/table/useTableStyle.js
@@ -1,154 +1,31 @@
-import { onMounted, onBeforeUnmount, computed, nextTick, ref, reactive, watch } from 'vue';
-import { isString, isFunction, isPlainObject, throttle, debounce, isArray } from 'lodash-es';
-import { ResizeObserver } from '@juggle/resize-observer';
+import { computed, ref, reactive } from 'vue';
+import {
+    isString,
+    isFunction,
+    isPlainObject,
+    throttle,
+    isArray,
+} from 'lodash-es';
 import getPrefixCls from '../_util/getPrefixCls';
 import { getCellValue } from './helper';
+import useTableLayout from './useTableLayout';
 
 const prefixCls = getPrefixCls('table');
-
-/**
- * 更新列的宽度
- * 列最小宽度为80，如果设定了width则使用设定的width作为宽度，没有设定宽度的列平分剩余的宽度（容器的宽度减去已使用的宽度）
- */
-function useTableLayout({ props, wrapperRef, headerWrapperRef, bodyWrapperRef, columns }) {
-    const widthList = ref([]);
-    const heightList = ref([]);
-    const bodyWidth = ref(0);
-    const isScrollX = ref(false);
-    const isScrollY = ref(false);
-    const wrapperHeight = ref(0);
-    const headerHeight = ref(0);
-    const bodyHeight = ref(0);
-
-    const handlerHeight = () => {
-        // 需要在宽度分配完，重新渲染后，此时table已经按照期望正常渲染，此时的高度才是最终高度
-        const $wrapper = wrapperRef.value;
-        const $bodyWrapper = bodyWrapperRef.value;
-        if ($wrapper && $bodyWrapper) {
-            const bodyWrapperHeight = $bodyWrapper.offsetHeight;
-            if (props.height) {
-                const $headerWrapper = props.showHeader ? headerWrapperRef.value : { offsetHeight: 0 };
-                const headerWrapperHeight = $headerWrapper.offsetHeight;
-                // 减去wrapperRef的border-bottom
-                const remainBodyHeight = props.height - headerWrapperHeight - 1;
-                bodyHeight.value = remainBodyHeight;
-                if (remainBodyHeight < bodyWrapperHeight) {
-                    isScrollY.value = true;
-                }
-                headerHeight.value = headerWrapperHeight;
-            }
-            wrapperHeight.value = $wrapper.offsetHeight;
-            heightList.value = Array.from($bodyWrapper.querySelectorAll('tbody tr')).map(($tr) => $tr.offsetHeight);
-        }
-    };
-
-    const handlerWidth = () => {
-        if (wrapperRef.value) {
-            const $wrapper = wrapperRef.value;
-            const _wrapperWidth = $wrapper.offsetWidth;
-            const wrapperWidth = props.bordered ? _wrapperWidth - 2 : _wrapperWidth;
-            let bodyMinWidth = 0;
-            const min = 80;
-            const _widthList = [];
-            columns.value.forEach((column) => {
-                const widthObj = {
-                    id: column.id,
-                };
-                const width = column.props.width;
-                const minWidth = column.props.minWidth;
-                if (width) {
-                    // 用户设置的宽度优先级最高
-                    bodyMinWidth += width;
-                    widthObj.width = width;
-                } else if (minWidth) {
-                    // minWidth次之
-                    bodyMinWidth += minWidth;
-                    widthObj.minWidth = minWidth;
-                } else if (column.props.type === 'selection' || column.props.type === 'expand') {
-                    // 展开和选择列固定为80（如果没有设置宽度的话）
-                    bodyMinWidth += min;
-                    widthObj.width = min;
-                } else {
-                    bodyMinWidth += min;
-                }
-                _widthList.push(widthObj);
-            });
-            widthList.value = _widthList;
-            const needAddWidthColumns = widthList.value.filter((column) => !column.width);
-            // 如果不够，则需要补宽度
-            if (bodyMinWidth < wrapperWidth) {
-                bodyWidth.value = wrapperWidth;
-                const surplus = (wrapperWidth - bodyMinWidth) % needAddWidthColumns.length;
-                const average = (wrapperWidth - bodyMinWidth - surplus) / needAddWidthColumns.length;
-                needAddWidthColumns.forEach((column, index) => {
-                    column.width = (column.minWidth || min) + (index === 0 ? average + surplus : average);
-                });
-            } else {
-                isScrollX.value = true;
-                bodyWidth.value = bodyMinWidth;
-                needAddWidthColumns.forEach((column) => {
-                    column.width = column.minWidth || min;
-                });
-            }
-        }
-    };
-
-    // 检测Table宽度变化
-    const ro = new ResizeObserver(
-        debounce(() => {
-            nextTick(() => {
-                handlerWidth();
-            });
-        }, 100),
-    );
-
-    watch(wrapperRef, ($wrapper) => {
-        if ($wrapper) {
-            ro.observe($wrapper);
-        }
-    });
-
-    onBeforeUnmount(() => {
-        ro.disconnect();
-    });
-
-    watch([columns, () => props.bordered, wrapperRef], handlerWidth);
-
-    // 当宽度计算出来，table渲染后，这是height才固定
-    // 第一层nextTick避免在width计算出来后无意义的计算
-    nextTick(() => {
-        watch(
-            [widthList, () => props.height, () => props.showHeader, wrapperRef, bodyWrapperRef, headerWrapperRef],
-            () => {
-                // 假如宽度发生变化，则需要等待渲染后再计算
-                nextTick(handlerHeight);
-            },
-            {
-                immediate: true,
-            },
-        );
-    });
-
-    return {
-        widthList,
-        heightList,
-        wrapperHeight,
-        headerHeight,
-        bodyWidth,
-        bodyHeight,
-        isScrollX,
-        isScrollY,
-    };
-}
 
 export default ({ props, columns, expandColumn, isExpandOpened }) => {
     const wrapperRef = ref(null);
     const headerWrapperRef = ref(null);
     const bodyWrapperRef = ref(null);
-    const fixedBodyWrapperRef = reactive([]);
-    const fixeHeaderWrapperRef = reactive([]);
 
-    const wrapperClass = computed(() => [prefixCls, props.bordered && 'is-bordered', props.size && `is-size-${props.size}`].filter(Boolean).join(' '));
+    const wrapperClass = computed(() =>
+        [
+            prefixCls,
+            props.bordered && 'is-bordered',
+            props.size && `is-size-${props.size}`,
+        ]
+            .filter(Boolean)
+            .join(' '),
+    );
 
     const layout = useTableLayout({
         wrapperRef,
@@ -158,9 +35,33 @@ export default ({ props, columns, expandColumn, isExpandOpened }) => {
         columns,
     });
 
+    const scrollState = reactive({
+        x: 'left',
+        y: false,
+    });
+
     const wrapperStyle = {};
 
+    const headerWrapperClass = computed(() => {
+        const arr = [`${prefixCls}-header-wrapper`];
+        if (scrollState.y) {
+            arr.push('is-scrolling-y');
+        }
+        if (scrollState.x) {
+            arr.push(`is-scrolling-x-${scrollState.x}`);
+        }
+        return arr;
+    });
+
     const headerWrapperStyle = {};
+
+    const bodyWrapperClass = computed(() => {
+        const arr = [`${prefixCls}-body-wrapper`];
+        if (scrollState.x) {
+            arr.push(`is-scrolling-x-${scrollState.x}`);
+        }
+        return arr;
+    });
 
     const bodyWrapperStyle = computed(() => {
         const style = {};
@@ -225,24 +126,39 @@ export default ({ props, columns, expandColumn, isExpandOpened }) => {
         }
     };
 
-    const getColClassName = ({ row, column, rowIndex, columnIndex }) => {
-        const colClassName = column.props.colClassName;
-        const cellValue = getCellValue(row, column);
-        if (isString(colClassName)) {
-            return colClassName;
+    const getCellClass = ({ column }) => {
+        const arr = [`${prefixCls}-cell`, column.id];
+        if (column.fixLeft) {
+            arr.push(`${prefixCls}-fixed-left`);
         }
-        if (isFunction(colClassName)) {
-            return colClassName({
-                row,
-                column,
-                rowIndex,
-                columnIndex,
-                cellValue,
-            });
+        if (column.fixRight) {
+            arr.push(`${prefixCls}-fixed-right`);
         }
+        return arr;
     };
 
-    const getColStyle = ({ row, column, rowIndex, columnIndex }) => {
+    const getCustionCellClass = ({ row, column, rowIndex, columnIndex }) => {
+        const colClassName = column.props.colClassName;
+        const cellValue = getCellValue(row, column);
+        const arr = [];
+        if (isString(colClassName)) {
+            arr.push(colClassName);
+        }
+        if (isFunction(colClassName)) {
+            arr.push(
+                colClassName({
+                    row,
+                    column,
+                    rowIndex,
+                    columnIndex,
+                    cellValue,
+                }) || '',
+            );
+        }
+        return arr;
+    };
+
+    const getCustomCellStyle = ({ row, column, rowIndex, columnIndex }) => {
         const cellValue = getCellValue(row, column);
         const colStyle = column.props.colStyle;
         const align = column.props.align;
@@ -289,45 +205,25 @@ export default ({ props, columns, expandColumn, isExpandOpened }) => {
         };
     };
 
-    const setScrollClassByEl = (el, className) => {
-        if (!el) return;
-        const classList = Array.from(el.classList).filter((item) => !item.startsWith('is-scrolling'));
-        classList.push(className);
-        el.className = classList.join(' ');
-    };
-
     // 同步两个table的位移
     const syncPosition = throttle(() => {
         const $bodyWrapper = bodyWrapperRef.value;
         if (!$bodyWrapper) return;
-        const { scrollLeft, scrollTop, offsetWidth, scrollWidth } = $bodyWrapper;
+        const { scrollLeft, scrollTop, offsetWidth, scrollWidth } =
+            $bodyWrapper;
         const $headerWrapper = headerWrapperRef.value;
         if ($headerWrapper) {
             $headerWrapper.scrollLeft = scrollLeft;
         }
-        const $fixedBodyWrapper = fixedBodyWrapperRef;
-        if ($fixedBodyWrapper.length) {
-            $fixedBodyWrapper.forEach((item) => {
-                item.scrollTop = scrollTop;
-            });
-        }
         const maxScrollLeftPosition = scrollWidth - offsetWidth - 1;
-        const isScrollX = layout.isScrollX.value;
         if (scrollLeft >= maxScrollLeftPosition) {
-            setScrollClassByEl($bodyWrapper, isScrollX ? 'is-scrolling-right' : '');
+            scrollState.x = 'right';
         } else if (scrollLeft === 0) {
-            setScrollClassByEl($bodyWrapper, isScrollX ? 'is-scrolling-left' : '');
+            scrollState.x = 'left';
         } else {
-            setScrollClassByEl($bodyWrapper, isScrollX ? 'is-scrolling-middle' : '');
+            scrollState.x = 'middle';
         }
-        const isScrollY = layout.isScrollY.value;
-        setScrollClassByEl($headerWrapper, isScrollY && scrollTop > 0 ? 'is-scrolling' : '');
-        const $fixeHeaderWrapper = fixeHeaderWrapperRef;
-        if ($fixeHeaderWrapper.length) {
-            $fixeHeaderWrapper.forEach((item) => {
-                setScrollClassByEl(item, isScrollY && scrollTop > 0 ? 'is-scrolling' : '');
-            });
-        }
+        scrollState.y = scrollTop > 0;
     }, 10);
 
     const handleHeaderMousewheel = (e, data) => {
@@ -338,39 +234,27 @@ export default ({ props, columns, expandColumn, isExpandOpened }) => {
         }
     };
 
-    const handleFixedMousewheel = (e, data) => {
-        const { pixelX, pixelY } = data;
-        if (Math.abs(pixelY) >= Math.abs(pixelX)) {
-            e.preventDefault();
-            bodyWrapperRef.value.scrollTop += data.pixelY;
-        }
-    };
-
-    onMounted(() => {
-        setScrollClassByEl(bodyWrapperRef.value, 'is-scrolling-left');
-    });
-
     return {
+        prefixCls,
         wrapperRef,
         headerWrapperRef,
         bodyWrapperRef,
-        getRowClassName,
-        getRowStyle,
-        getColClassName,
-        getColStyle,
-        wrapperClass,
-        prefixCls,
         layout,
+        wrapperClass,
         wrapperStyle,
         headerWrapperStyle,
         bodyWrapperStyle,
         headerStyle,
         bodyStyle,
         getCellSpan,
+        getRowClassName,
+        getRowStyle,
+        getCellClass,
+        getCustionCellClass,
+        getCustomCellStyle,
         syncPosition,
-        fixedBodyWrapperRef,
-        fixeHeaderWrapperRef,
         handleHeaderMousewheel,
-        handleFixedMousewheel,
+        headerWrapperClass,
+        bodyWrapperClass,
     };
 };