+++ components/version/index.js
@@ -1,3 +1,3 @@
-const version = '0.0.3';
+const version = '0.0.7';
 
 export default version;