+++ docs/.vitepress/components/cascader/common.vue
@@ -1,13 +1,13 @@
 <template>
     <div class="text-tips">默认单选：</div>
     <FCascader       
-        v-model="base.value"
+        v-model="base.value1"
         :options="base.options"
         @change="base.handleChange">
     </FCascader>
     <div class="text-tips">hover单选：</div>
     <FCascader       
-        v-model="base.value"
+        v-model="base.value2"
         :options="base.options" 
         :nodeConfig="{ expandTrigger: 'hover' }"
         @change="base.handleChange">
@@ -16,31 +16,6 @@
     <FCascader       
         v-model="base.valueEmitPath"
         :options="base.options"
-        :showAllLevels="false"
-        :clearable="true"
-        :nodeConfig="{ emitPath: true }"
-        @change="base.handleChange">
-    </FCascader>
-    <div class="text-tips">默认多选：</div>
-    <FCascader       
-        v-model="base.multiValue"
-        :options="base.options"
-        :multiple="true"
-        @change="base.handleChange">
-    </FCascader>
-    <div class="text-tips">hover多选：</div>
-    <FCascader       
-        v-model="base.multiValue"
-        :options="base.options" 
-        :multiple="true"
-        :nodeConfig="{ expandTrigger: 'hover' }"
-        @change="base.handleChange">
-    </FCascader>
-    <div class="text-tips">多选不展示路径：</div>
-    <FCascader       
-        v-model="base.multiValueEmitPath"
-        :options="base.options"
-        :multiple="true"
         :showAllLevels="false"
         :clearable="true"
         :nodeConfig="{ emitPath: true }"
@@ -233,10 +208,9 @@
     // 基础用法
     const useBase = () => {
         const state = reactive({
-            value: '',
+            value1: '',
+            value2: '',
             valueEmitPath: [],
-            multiValue: [],
-            multiValueEmitPath: []
         })
 
         return {
@@ -262,8 +236,6 @@
 .fes-cascader {
     width: 200px;
 }
-</style>
-<style lang="less" scoped>
 .text-tips {
     margin-top: 10px;
 }