+++ docs/.vitepress/components/message/closeTime.vue
@@ -7,7 +7,7 @@
 </template>
 
 <script>
-import { FMessage } from 'fes-design';
+import { FMessage } from '@fesjs/fes-design';
 
 export default {
     setup() {