+++ docs/.vitepress/components/pagination/combination.vue
@@ -1,3 +1,3 @@
 <template>
-    <FPagination :total-count="20" show-size-changer show-quick-jumper show-total></FPagination>
+    <FPagination :total-count="1000" show-size-changer show-quick-jumper show-total></FPagination>
 </template>