+++ docs/.vitepress/components/tooltip/confirm.vue
@@ -14,7 +14,7 @@
 </template>
 
 <script>
-import { FMessage } from 'fes-design';
+import { FMessage } from '@fesjs/fes-design';
 
 export default {
     setup() {